Moved all implementation inside the API class
